fix issue with mv_contains function

+++ src/main/java/org/apache/druid/extensions/aggregators/multivalue/BooleanOrBufferAggregator.java
@@ -41,32 +41,32 @@ public class BooleanOrBufferAggregator implements BufferAggregator
   @Override
   public void init(ByteBuffer buf, int position)
   {
-    buf.putLong(position, 0);
+    buf.putInt(position, 0);
   }
 
   @Override
   public void aggregate(ByteBuffer buf, int position)
   {
-    buf.putLong(position, Math.max(buf.getLong(position), selector.getObject() ? 1 : 0));
+    buf.putInt(position, Math.max(buf.getInt(position), selector.getObject() ? 1 : 0));
   }
 
   @Nullable
   @Override
   public Object get(ByteBuffer buf, int position)
   {
-    return buf.getLong();
+    return buf.getInt();
   }
 
   @Override
   public float getFloat(ByteBuffer buf, int position)
   {
-    return buf.getLong();
+    return buf.getInt();
   }
 
   @Override
   public long getLong(ByteBuffer buf, int position)
   {
-    return buf.getLong();
+    return buf.getInt();
   }
 
   @Override
